Add form_assign plugin
This is useful when you want to assign attachment from form params but
don't want to create a form object for that.

  attacher.form_assign("file" => raw_file)
  attacher.file #=> #<Shrine::UploadedFile>

It works with data_uri and remote_url plugins as well:

  attacher.form_assign("file_remote_url" => "https://example.com/file")
  attacher.file #=> #<Shrine::UploadedFile>

By default it returns params with cached file data, in format which can
be assigned again. This acts as sort of normalization of different
attached file sources.

  attacher.form_assign("file" => raw_file, "foo" => "bar")
  #=> { "file" => '{"id":"...","storage":"...","metadata":{...}}', "foo" => "bar" }

You can also have it return attributes for persisting:

  attacher.form_assign({ "file" => raw_file, "foo" => "bar" }, result: :attributes)
  #=> { "file_data" => '{"id":"...","storage":"...","metadata":{...}}', "foo" => "bar" }

Internally this loads the model plugin, but only into a disposable
Shrine subclass, as to not make any side effects if people are using
Shrine::Attachment with immutable entities.