Change #create_derivatives to pass arguments to #process_derivatives
I think it will be much more common for people to want to pass arguments
to the processor than to the Shrine uploader. I expect on the uploader
side people will mainly want to configure upload storage and options,
and both can be configured on the class level.

It's also easier to remember that Attacher#create_derivatives and
Attacher#process_derivatives both have the same method signature, rather
than that Attacher#create_derivatives is half #process_derivatives and
half #upload_derivatives.