Add #<name>_derivatives! wrapper around Attacher#create_derivatives
To generate derivatives in the controller, the user currently needs to
call Attacher#create_derivatives directly:

  photo = Photo.new
  photo.image = file
  photo.image_attacher.create_derivatives(:thumbnails)
  photo.image_derivatives #=> { small: ..., medium: ..., large: ... }

I don't think this is easy to remember, especially considering that
Attacher#create_derivatives is the API that we expect most people will
use. Accessing the attacher kind of looks like breakage of the Law of
Demeter, where we need to call a method on a seemingly internal object
(people who want to use Shrine on a simple level probably don't want to
know about Shrine::Attacher).

To make this main API more convenient, we add a new
`#<name>_derivatives!` method to the attachment module, which calls
Attacher#create_derivatives. Now our previous examples looks like this:

  photo = Photo.new
  photo.image = file
  photo.image_derivatives!(:thumbnails)
  photo.image_derivatives #=> { small: ..., medium: ..., large: ... }

I think this is a big improvement in convenience. It's also nice that
there is this link between `#<name>_derivatives!` and
`#<name>_derivatives`, so people will remember more easily that the
latter method exists.

The idea of a bang method doing work was taken from the Dragonfly gem.
Also, the ImageProcessing gem does work when a bang method is called.