Improve default_storage plugin
Currently the dynamic storage block accepts record & name arguments. I
think this is limiting and coupled to entity/model plugin.

So we change storage block not to receive any arguments, and evaluate it
in context of the Attacher instance. We also add Attacher.default_cache
and Attacher.default_store for alternative way of configuring the
default storage, which communicates that the storage block is executed
in context of the Attacher instance.

We deprecate record & name arguments to the storage block.