Add Attacher#file_data as lighter Attacher#data
When we're using backgrounding, we pass current attached file data to
the background job, so that we can later assert that the attachment
hasn't changed while we do our expensive work. So far we were
recommending just sending Attacher#data, which will contain the needed
file data.

However, Attacher#data can also contain a lot of stuff we *don't* need
for Attacher.retrieve, such as metadata and derivatives data. This can
make it more difficult to look at our background job payload, because it
can be bloated with data we don't need. I noticed this when working on
shrine-transloadit, where I already had some derivatives and lots of
metadata.

So, we add Attacher#file_data to atomic_helpers plugin, which returns
only main attached file data, without derivatives or metadata. This
should then be passed into the background job.

We also rename the :data option in Attacher.retrieve to :file, to make
the intent more clear; we're passing what we expect to be the current
*file*, so there is no need to use a generic "data" term here.