Add mirroring plugin
The mirroring plugin adds the ability to replicate uploads and deletes
to other storages. This is useful for implementing a backup storage, or
or when migrating files from one storage to the other.

  plugin :mirroring, mirror: { store: :other_store }

With the above settings, any uploads and deletes to :store will also be
mirrored to :other_store. It's also possible to move mirroring into a
background job.