Handle hashes in column plugin
In ROM we don't have a way to determine whether an entity attribute will
return a Hash or not. So in the rom plugin we'd have to override
Attacher#deserialize_column to handle that. I think that it's better if
Shrine already handles this for us.

An added benefit is that we can call #to_hash on the resulting hash,
which will convert Sequel's pg_json objects into pure hashes. This isn't
strictly necessary, I just want that persistence-specific objects are
not leaking outside.