Remove references to ActiveRecord::Base.raise_in_transactional_callbacks
In Active Record 6.0 this setting doesn't exist anymore, in 5.x it
defaults to true, and AFAIK in version 4.x you get a warning when an
exception was raised (maybe in 3.x as well).

While Shrine technically supports Rails 3.2, I don't want to be
confusing people on newer versions of Rails. So we remove references to
ActiveRecord::Base.raise_in_transactional_callbacks.