Remove attacher_options plugin
With the recent backgrounding change, it's not possible anymore to use
attacher_options plugin to pass promote options to the background
block.

This doesn't make attacher_options useful enough to keep around. If
people want to pass options to attacher operations called in attachment
flow, they will need to drop to the attacher level.