Add support for default derivatives processors
In the most common case the user will have one derivatives processor for
a given uploader:

  class ImageUploader < Shrine
    Attacher.derivatives_processor :thumbnails do |original|
      # ...
    end
  end

  ...

  photo.image_derivatives!(:thumbnails)

In this case it can be convenient if the user doesn't have to name the
processor. So we add support for default derivatives processors:

  class ImageUploader < Shrine
    Attacher.derivatives_processor do |original|
      # ...
    end
  end

  ...

  photo.image_derivatives!

This is especially useful when generating derivatives in a background
job. In this case there be one worker that will handle multiple
uploaders, so if they do different kinds of processing the user would
have to handle that:

  class ImageUploader < Shrine
    Attacher.derivatives_processor(:thumbnails) { ... }
  end

  class VideoUploader < Shrine
    Attacher.derivatives_processor(:transcoded) { ... }
  end

  ...

  class PromoteJob < ActiveJob::Base
    def perform(record, name, file_data)
      attacher = Shrine::Attacher.retrieve(model: record, name: name, file: file_data)

      case attacher.shrine_class
      when ImageUploader then attacher.create_derivatives(:thumbnails)
      when VideoUploader then attacher.create_derivatives(:transcoded)
      end

      attacher.atomic_promote
    end
  end

With default processors the background job doesn't need to know for
which uploader it's performing the processing:

  class ImageUploader < Shrine
    Attacher.derivatives_processor { ... }
  end

  class VideoUploader < Shrine
    Attacher.derivatives_processor { ... }
  end

  ...

  class PromoteJob < ActiveJob::Base
    def perform(record, name, file_data)
      attacher = Shrine::Attacher.retrieve(model: record, name: name, file: file_data)
      attacher.create_derivatives # calls the default processor
      attacher.atomic_promote
    end
  end

To handle uploaders which don't have derivatives processor defined, we
make Attacher#create_derivatives a no-op in this case.