Accept only File/Tempfile object as derivation result
It adds unnecessary complication to accept paths as well. So we limit it
to opened file objects.

We also make sure to re-open the file only if it's a Tempfile object.
With Tempfiles this is useful because you'll usually first create an
empty one, then populate it possibly with an external program, in which
case the file descriptor needs to be refreshed.

With File objects that shouldn't be the case. Also, it's possible the
user used some special open modes that we would cancel by re-opening.
So, we still make sure binary mode is used, but we don't re-open.