Allow uploading files larger than 50 GB
This was possible in Shrine 2.x when Aws::S3::Object#upload_file was
used, as in this case the correct multipart part size was automatically
determined.

However, Aws::S3::Object#upload_stream doesn't know the total size of
the content, so it uses minimum part size by default. This doesn't work
for files larger 50 GB, as we'd be hitting AWS S3's limit of 10,000
parts.

So, when the file is larger than 50 GB, we make sure to use a large
enough part size. We cannot really test this, as #upload_stream will
block if we don't give it enough content, and we cannot really use 50 GB
files in our tests.