Update recommended backgrounding setup
Active Job 6.0 cannot yet serialize class/module names. I've seen that
functionality on Rails master, so I suppose it will be supported in
Rails 6.1, but it's not yet supported on 6.0. In fact, Active Job will
currently raise an exception if you pass it a class object, it won't
even try to call #to_s on it like Sidekiq will.

So, we update examples to also serialize and deserialize classes. While
we're here, we also update examples to use Sidekiq instead of Active
Job, and handle serialization of records. Yes, Active Job knows how to
dump and load Active Record objects, but we want our examples to be
generic and work with any web framework (rather than treating Rails
users as first class citizens).