Add Attachment[] shorthand for Attachment.new
This adds one more way of creating an attachment module:

  class Photo
    include Shrine::Attachment[:image]
  end

I think it's nice to allow people to choose between multiple syntaxes.
I personally wanted to use the square brackets syntax after working with
ROM.

With this we also remove mentions of the Shrine[] syntax to discourage
its use, as it's too implicit. It will likely get deprecated later and
removed in Shrine 4.