Add back ability to call #delete_prefixed with trailing slash
This was original part of https://github.com/shrinerb/shrine/pull/413,
but I've removed it because I wanted only one way to call
Storage#delete_prefixed.

However, some people prefer the syntax with a trailing slash, and I can
see the arguments behind it, so we add back support for that. Active
Storage uses this syntax as well, so it will be familiar to users
transitioning from there.

Co-authored-by: Jonathan Rochkind <jrochkind@sciencehistory.org>