Move setting :action to :store to Attacher#promote
If this is done on the Attacher#promote_cached level, then it won't be
passed with backgrounding, because in that case Attacher#promote is
called directly.

It's important that the behaviour is consistent when backgrounding is
used. It's also important for the processing/versions plugin
compatibility, which won't get triggered otherwise.