Switch from highlight.js to atom/highlights
Highlight.js sometimes doesn't tokenize ruby code correctly (e.g. it
doesn't detect constants). This is because it implements its own
regexes for parsing code, which is bound to have bugs. Prism seems to
have better regexes in general, but it still fails for certain cases.

There is actually a robust canonical Ruby grammar, which lives inside
the TextMate organization, and is used by GitHub, Atom and VScode.
Atom's "highlights" package uses this grammar, and it's really easy to
use, so we switch to it, and add a custom theme replicating zenburn.