Add Attacher.derivatives shorthand
Now we can shorten

  Attacher.derivatives_processor do |original|
    # ...
  end

to just

  Attacher.derivatives do |original|
  end

I think this helps make the API more streamlined, as that's the main
part, and it's now quicker to type and easier to remember.

We will still keep Attacher.derivatives_processor, as
Attacher.derivatives is just a shorthand, but I don't see any reason to
show Attacher.derivatives_processor in the documentation.