Propagate exceptions on loading ruby-vips in store_dimensions plugin
Loading the ruby-vips library can fail with errors other than LoadError
when the libvips package isn't available, and those exceptions *are* a
descendant of StandardError, so they would be rescued by the rescue
statement. This would cause an "uninitialized constant Vips::Error"
exception to be raised.

To fix this, we make sure to only rescue exceptions that happen during
analyzing, not while loading the library.