Add :storage option to Attacher#create_derivatives
Ability to specify destination storage is common enough that we want to
make it as convenient as possible. Since the :storage option for
Attacher#upload_derivatives is separate from other uploader options, we
modify Attacher#create_derivatives to accept it and forward it to
Attacher#add_derivatives.

Now, instead of

  files = photo.image_attacher.process_derivatives
  photo.image_attacher.add_derivatives(files, storage: :cache)

People can just write

  photo.image_derivatives!(storage: :cache)