Always return symbols in Attacher#cache_key & Attacher#store_key
 Rubyists sometimes treat strings and symbols indifferently when doing
 interpolation, and this might be the case when providing cache and
 store key for the attacher. If this is the case, methods like
 Attacher#cached? and Attacher#stored? will not work correctly as they
 expect storage keys to be symbols.

At first I thought about modifying Attacher#uploaded? to handle symbol
keys instead. But then I realized that we might potentially need to
handle these cases in other places instead, and that it's better to
coerce user input as early on as possible.

We cannot do it in Attacher#initialize as that would break
default_storage plugin when dynamic block is used, so we do it in
Attacher#cache_key and Attacher#store_key.