Add section about URL fallbacks
It has been asked how to create a fallback for missing derivative to
show the original URL. When versions plugin was used, we also added the
feature to fall back to another derivative URL, as per a request. The
demo app shows a third fallback strategy, which is to use on-the-fly
processing.

So, we add a new section in the "File Processing" guide which shows
these three approaches.