Add ignore_assign_current plugin
When coupled with a SPA, some users want to treat the attachment
attribute as a permanent resource attribute, meaning that the API will
always return attached file data (whether it's cached or stored) and the
frontend will send the attached file data back on resource updates.

This was possible with Shrine 2.x, as then the attacher would simply
skip any non-cached files. This changed in Shrine 3.0, where attempting
to assign a non-cached file will now raise an exception, to let you know
the operation didn't succeed.

However, for some people this feature was convenient as it meant less
boilerplate when building a SPA + API, so we add it back in form of a
plugin. It makes the following assignment a no-op:

  photo.image #=> #<Shrine::UploadedFile id="foo" storage=:store ...>
  photo.image = '{"id":"foo","storage":"store",...}' # no-op

The plugin implementation is not as efficient as it could be; it will
parse the cached attached file JSON data twice on assignment. However,
avoiding the double-parse would require some non-trivial changes to the
base API which I don't think are worth it at the moment.

Fixes #420