Show uploader-specific on-the-fly processing
It's confusing when eager processing is in the uploader, but on-the-fly
processing is outside. So we change examples to show defining
derivations inside uploaders. We configure the secret key globally in
the initializer so that it's not repeated for each uploader.

With this approach users will need to mount a derivation endpoint for
each uploader they have, but I think that's ok since I expect most
people will have only one uploader.