Fix :derivative inconsistency when promoting derivatives
When derivatives are being promoted, because we're not calling
`#derivative_path`, the :derivative name would always be an array,
whether it's top level or nested. This is inconsistent with regular
derivatives creation, where top level :derivative names are just
symbols.

We fix that by moving the `#derivative_path` call into
Attacher#upload_derivative, so that we ensure consistent naming for any
operation that calls this method, which is currently
Attacher#upload_derivatives and Attacher#promote_derivatives.

This means the user cannot use a top-level array for the path when
calling Attacher#upload_derivative if they wanted to, but I think that's
a reasonable trade-off for ensuring consistent naming.