Add allow_reassign plugin (#432)
When coupled with a SPA, some users want to treat the attachment
attribute as a permanent resource attribute, meaning that the API will
always return attached file data (whether it's cached or stored) and the
frontend will send the attached file data back on resource updates.

This was possible with Shrine 2.x, as then the attacher would simply
skip any non-cached files. This changed in Shrine 3.0, where attempting
to assign a non-cached file will now raise an exception, to let you know
the operation didn't succeed.

However, for some people this feature was convenient as it meant less
boilerplate when building a SPA + API, so we add it back in form of a
plugin. It makes the following assignment a no-op:

  photo.image #=> #<Shrine::UploadedFile id="foo" storage=:store ...>
  photo.image = '{"id":"foo","storage":"store",...}' # no-op