Make remove_invalid plugin work with derivatives
There were three plugins overriding Attacher#change:

  * derivatives
  * validation
  * remove_invalid

If we load the plugins in the following order:

  Shrine.plugin :derivatives
  Shrine.plugin :remove_invalid

That would produce the Attacher#change method with the following
pseudocode:

  def change(file)
    @previous = dup
    set(file)
    set_derivatives({})
    validate
    remove_invalid
    file
  end

The only thing here that didn't work was that remove_invalid would only
assign back the previous *main* file in case of validation errors, not
the derivatives, so we fix that by using Attacher#load_data.

However, if we load the plugins in the opposite order:

  Shrine.plugin :remove_invalid
  Shrine.plugin :derivatives

This would produce Attacher#change as follows:

  def change(file)
    @previous = dup
    set(file)
    validate
    remove_invalid
    set_derivatives({})
    file
  end

The above would not work correctly, because at the time of the
validation the derivatives of the previous attachment aren't yet
cleared. This causes remove_invalid to falsely destroy derivatives of
the previous attachment while reverting.

We fix that by removing validation and remove_invalid behaviour from
Attacher#change, and keeping only derivatives. We move automatic
validation to Attacher#attach and Attacher#attach_cached, as that way
it's triggered both with caching on and off in the model plugin. We also
move remove_invalid behaviour into Attacher#validate, which makes more
sense. This fixes the load order bug.