Simplify @previous handling
Originally, the "@previous" variable was holding the uploaded file
directly. In order to distinguish between when file changed with
"@previous" being nil and when there was no change, we introduced
checking for instance variable presence.

However, once we switched to "@previous" holding an attacher copy (even
when there was no previous file), we can now consider nil value when
the attacher hasn't changed, which simplifies code.