Merge allow_reassign plugin into base
Prior to Shrine 3.0, it was possible to assign a stored file, which
would be a no-op. This enabled people to treat the attachment attribute
as a persistent attribute, where the current attached file data can be
assigned back. This functionality could simplify integration with a SPA. [1]

In Shrine 3.0, assigning a stored file changed to raising an exception,
to alert the user that the assignment failed. However, this also
disallowed assigning the current stored file. So we added back the
functionality to assign back the current file.

We initially added this functionality as a separate allow_reassign
plugin. However, I think this behaviour deserves to be in the base, and
it also simplifies some things such as rack_file plugin integration.

This implementation adds the additional overhead of parsing cached JSON
data twice on assignment, but I think that's a good trade-off for its
readability.

[1]: https://github.com/shrinerb/shrine/issues/420#issuecomment-556291632