Make expectations with empty hash arguments work on Ruby 2.7
In Ruby 2.7, passing a double-splatted empty hash is the same as passing
no keyword arguments. So, when upload options are empty, method such as
Shrine#extract_metadata get without a second argument, which fails the
expectation because we expect it to get called with an empty hash.

Using `**{}` would work on Ruby 2.5 and older and Ruby 2.7, because on
Ruby 2.7 it would be the same as if we passed on arguments, while on
Ruby 2.5 and older it would be the same as if we passed a hash. However,
this wouldn't work on Ruby 2.6. Like Ruby 2.7, Ruby 2.6 treats `**{}` as
if no argument was passed, but unlike Ruby 2.7, double-splatting a local
variable that holds an empty hash (i.e. `**empty_hash`) *does* pass it
as an empty hash. So Mocha expectation fails, because it expects the
method will be called with an empty hash.

To work around this, we assign an empty hash to a variable and
double-splat that. That way it works regardless of how Ruby passes empty
keyword arguments.