Work around apparent StringIO bug in ruby 2.7
Confirmed the test added here does fail in 2.7 without the code change, and passes with the code change.

Also confirmed this code change is what was needed for my actual app, which uses Storage::Memory in tests, to pass it's own tests in ruby 2.7. (Although my actual app is using shrine 2.x still, and made equivalent change in shrine 2.x Storage::Memory).

This appears to be a ruby bug in 2.7.0.

https://bugs.ruby-lang.org/issues/16497

https://www.reddit.com/r/ruby/comments/emd6q4/is_this_a_stringio_bug_in_ruby_270/

Yes, in this test we still need to supply an encoding to the StringIO in the test, also because of the ruby 2.7.0 bug. However, even doing that, the test fails without the code change here, and passes with it. And my actual app needed no other changes than the change here to pass again.