Add tests that Memory#upload and Memory#open accept kwargs
This is important when memory storage is used for tests, in which case
there might be some upload/download options passed by the application
code.

We also want to make sure there aren't any kwargs warnings being
triggered on Ruby 2.7, which could have been the case considering we
were using "*" instead of "**" (though it turns out it's not).