Add back S3 client-side encryption
We broke client-side encryption in Shrine 3.0 when we removed the
additional HEAD request in S3#open and started building the get_object
request directly. Furthermore, aws-sdk-s3 broke initializing
Aws::S3::Bucket with an Aws::S3::Encryption::Client object in one of the
recent versions.

We add back proper support for client-side encryption, as a separate
module that adds encryption-specific logic.

Closes #348