Add :create_on_promote option for derivatives
When you're attaching files on multiple places, it can be cumbersome to
remember to create derivatives in each place. This is not a problem when
backgrounding is used, as there derivatives are created in a single
place (the worker), it's mainly in a synchronous setup.

This is how CarrierWave and Paperclip behave, and it's also how Shrine's
versions plugin behaved in Shrine 2.x. However, with those solutions you
didn't have control over when the derivatives will be generated, or to
generate derivatives in multiple tiers. Here the auto-generation is
opt-in, so we still keep the flexibility and explicitness should the
developer want it.

We also make documentation updates:

* the "Getting Started" guide shows the simplest setup, which is with
  :create_on_promote

* the "File Processing" guide shows the expanded setup with validation,
  as there subsequent sections already rely on :create_on_promote *not*
  being enabled (showing creation flexibility)

* the "Derivatives" plugin guide shows the expanded setup without
  validation, as the goal of that guide is to explain how derivatives
  works.