Delete removed attachment if new file was attached right after removal
When the remove checkbox is ticked and a new file was attached, the
update issued would be as follows:

  post.update(remove_image: true, image: new_file)

Because Attacher#change just saves whichever was the previous file, the
"remove image" operation would save the original file, and then the
"attach" operation would overwrite the previous file again, this time
with nil because that was the current attached file.

Since this is a common scenario, we handle this scenario in the
remove_attachment plugin, where the second Attacher#change call will
keep the previous file pointing to the removed file.

Closes #473