Fix S3#open not working on aws-sdk-core 3.104+
The newest aws-sdk-core version changed the Seahorse client to return
response headers as the second block argument when streaming a GET
request. We're using this via the get_object operation in S3#open.

This is actually a feature I requested, since previously we've needed to
hack into aws-sdk-s3 to fetch the Content-Length header. We split the
S3#get method into two versions, to still handle aws-sdk-core versions
< 3.104.