Apply `version` query parameter in derivation_endpoint
When passing `:version` to `#derivation_url`, the option would get
correctly added to query parameters, and it would fulfill the role of
invalidating the cache, but it wasn't being forwarded to the derivation.

This meant that, if a person was using :upload option and was basing
:upload_location on :version, the upload location wouldn't change, and
possibly the cached file would be served.

We fix that by passing the `version` query parameter to the derivation.