Call Attacher#reload when reloading ActiveRecord/Sequel model
ActiveRecord::Base#reload and Sequel::Model#reload reload the column
values from the database, but they keep all other state. However, in
Shrine we throw away the whole attacher in this case. This may be
undesired in case the user had set some ad-hoc attacher state, such as
context or storage, and want it to be retained after model reload.

So we change sequel and activerecord plugins to call Attacher#reload
instead, which only reloads the file and derivatives, but keeps
everything else.

Closes #463