Support Aws::S3::EncryptionV2::Client
This is a new, more secure, client-side encryption implementation added
to newer versions of aws-sdk-s3.

Attempting to stream #get_object would trigger a deprecation warning, so
we download all content into a file first, and then read chunks from it.
This is a bit wasteful, since UploadedFile#download will write the
content to disk again, but it's a price to pay for respecting the
abstraction, and most people won't use client-side encryption.

Concerns #348