Redefine Column#type method directly in activerecord_test.rb
Setting instance variables on this internal structure doesn't work in
Rails 6.1.0.rc1, because that object is now frozen. So we choose a much
simpler approach that just monkey-patches the method the activerecord
plugin is checking. We do this on a copy of the column, because the
original column object is reused between tests.