Use plain attr readers & accessors in examples
Using imaginary `Entity()` and `Model()` class builder methods can be
confusing for newcomers, as it's something that was not introduced
before, and indicate the methods exist whereas they do not.

Instead we use plain attr readers and accessors. The examples won't
actually work just with these classes, because they assume the class
also has an #initialize method which accepts a hash of attributes. But I
think it's close enough.