Fixes error in mime type given to svgs by fastimage
The determine_mime_type plugin returns the incorrect mime type
for svgs when using the fastimage analyzer. This is because fastimage
is returning a symbol `:svg` as the type, and Shrine is just doing a
string interpolation to return `image/` plus whatever fastimage returns.

The correct mime type for svgs is `image/svg+xml`, which requires a
special case in this method since simple string interpolation won't
handle it. So I've added a return value for `image/svg+xml` for this
case.