S3 should use tagging_directive REPLACE on copy operation (#569)
It was always using metadata_directive: REPLACE; this does not apply to AWS tags applied to an S3 object. It should also use tagging_directive to apply to AWS tags as well.

This makes it so tags aren't copied when copying from (say) cache to storage. More significantly, it allows the caller to pass S3 `tagging` param in `upload_options`, and have this take effect, whether on a promotion or any other kind of `upload`, regardless of whether the S3 adapter decides to use a `copy_from` operation under the hood.

Without this change, whether `tagging` passed as `upload_options` actually works (as well as whether the destination winds up with copied tags from source) depends on whether the S3 adapter decides to use `copy_from`, which should be an implementation detail that does not effect outcomes.

More at https://discourse.shrinerb.com/t/gotcha-on-s3-upload-options-and-tags/559