Remove issue templates
I wanted to request help for bring the website search feature back, and
there was no appropriate way for me to do that. I could either open a
bug report in the library or open a discussion. Given this was a bug on
the website, I still wanted to open an issue, but the bug report
template wasn't appropriate for this.

This gem is not nearly popular enough that an issue template will bring
value, and the friction cost is not worth it to me.