Simplify Rack 3 support using `Rack::Headers`
The `Rack::Headers` subclass of `Hash` was added in Rack 3, and is meant
to help libraries support both Rack 2 and 3. I was concerned whether
returning `Rack::Headers` instead of `Hash` to the end user was
backwards compatible, but it seems that both Roda and Rack are doing it,
so it should be safe.

We remove Appraisal and for now let Rack 2 be tested implicitly with
Ruby 2.3, as Rack 3 requires Ruby 2.4+. But I don't think we need to be
testing Rack 2.x support, especially now that we've removed duplication,
I would be surprised if it breaks. With `Rack::Headers` we were able to
keep Ruby 2.3 support, as we're not using `Hash#transform_keys` anymore.

I also switched from rack-test_app, which seems to be unmaintained, back
to rack-test, which is the established library for testing Rack apps.
Ever since Jeremy Evans took over maintenance, it's gotten a lot better
and more flexible, so it's completely viable for Shrine's needs. Since
rack-test returns `Rack::Headers`, we were able to remove the header
constants.